Change ScrollParam in ScrollBarNode.cs from int to float

+++ Nodes/ComponentNode/ScrollBarNode.cs
@@ -35,9 +35,9 @@ public unsafe class ScrollBarNode : ComponentNode<AtkComponentScrollBar, AtkUldC
     /// <summary>
     /// Gets or sets the current scroll position, triggering the component to update.
     /// </summary>
-    public int ScrollPosition {
+    public float ScrollPosition {
         get => Component->ScrollPosition;
-        set => Component->SetScrollPosition(value);
+        set => Component->SetScrollPosition((int) value);
     }
 
     /// <summary>
